Docs: 회원 기능 분류
