+++ src/main/java/com/erp/domain/client/controller/ManagerClientController.java
@@ -2,7 +2,6 @@ package com.erp.domain.client.controller;
 
 import com.erp.domain.client.dto.request.AddClientBlacklist;
 import com.erp.domain.client.dto.response.ClientDetailResponse;
-import com.erp.domain.client.dto.response.ClientRentHistoryResponse;
 import com.erp.domain.client.dto.response.ClientSummaryResponse;
 import com.erp.domain.client.service.ManagerClientService;
 import jakarta.validation.Valid;
@@ -48,6 +47,12 @@ public class ManagerClientController {
     @PostMapping("/black/{clientId}")
     public ResponseEntity<Void> addBlacklist(@PathVariable Long clientId, @Valid @RequestBody AddClientBlacklist dto) {
         clientService.addBlacklist(clientId, dto);
+        return ResponseEntity.noContent().build();
+    }
+
+    @DeleteMapping("/black/{clientId}")
+    public ResponseEntity<Void> removeBlacklist(@PathVariable Long clientId) {
+        clientService.removeBlacklist(clientId);
         return ResponseEntity.noContent().build();
     }
 }
