+++ src/main/java/com/erp/domain/client/service/ClientService.java
@@ -200,17 +200,4 @@ public class ClientService {
         return LocalDate.of(year, month, day);
     }
 
-    @Transactional
-    public void removeBlacklist(Long clientId) {
-        Client client = clientRepository.findById(clientId)
-                .orElseThrow(() -> new CustomException(404, "해당 고객을 찾을 수 없습니다."));
-
-        boolean isBlack = client.getBlacklisted() != null && client.getBlacklisted();
-        if (!isBlack) {
-            throw new CustomException(400, "이미 블랙리스트 해제 상태인 고객입니다.");
-        }
-
-        client.setBlacklisted(false);
-        client.setBlacklistInfo(null);
-    }
 }
