+++ src/main/java/com/erp/domain/client/service/ManagerClientService.java
@@ -3,11 +3,9 @@ package com.erp.domain.client.service;
 import com.erp.domain.accident.repository.AccidentRepository;
 import com.erp.domain.client.dto.request.AddClientBlacklist;
 import com.erp.domain.client.dto.response.ClientDetailResponse;
-import com.erp.domain.client.dto.response.ClientRentHistoryResponse;
 import com.erp.domain.client.dto.response.ClientSummaryResponse;
 import com.erp.domain.client.entity.Client;
 import com.erp.domain.client.repository.ClientRepository;
-import com.erp.domain.rent.entity.Rent;
 import com.erp.domain.rent.repository.RentRepository;
 import com.erp.global.exception.CustomException;
 import lombok.AllArgsConstructor;
@@ -138,4 +136,18 @@ public class ManagerClientService {
 //                rent.getEndRentDateTime()
 //        );
 //    }
+
+    @Transactional
+    public void removeBlacklist(Long clientId) {
+        Client client = clientRepository.findById(clientId)
+                .orElseThrow(() -> new CustomException(404, "해당 고객을 찾을 수 없습니다."));
+
+        boolean isBlack = client.getBlacklisted() != null && client.getBlacklisted();
+        if (!isBlack) {
+            throw new CustomException(400, "이미 블랙리스트 해제 상태인 고객입니다.");
+        }
+
+        client.setBlacklisted(false);
+        client.setBlacklistInfo(null);
+    }
 }
